Public Documentation

Documentation for ClimaOcean.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

ClimaOcean

Diagnostics

ClimaOcean.Diagnostics.MixedLayerDepthField — Method

MixedLayerDepthField(grid, tracers, buoyancy_model; Δb = 3e-4, field_kw...)

Return a reduced Field{Center, Center, Nothing} that represents mixed layer depth for model, based on a buoyancy differential criterion. The mixed layer depth is defined as the depth $h$ for which

\[b(z=0) - b(z=-h) = Δb\]

This criterion is solved by integrating downwards and linearly interpolating to find h, assuming that $b$ decreases with depth.

Keyword arguments

Δb: Buoyancy differential used to calculate mixed layer depth

field_kw: Keyword arguments passed to Field.

Example

h = MixedLayerDepth(model)
compute!(h) # compute mixed layer depth