Private types and functions

Documentation for ClimaOcean.jl's internal interface.

ClimaOcean

Diagnostics

InitialConditions

DataWrangling

ClimaOcean.DataWrangling.continue_downwards! — Method

continue_downwards!(field, mask)

Continue downwards a field with missing values within mask. Cells where mask[i, k, k] == false will be preserved.

ClimaOcean.DataWrangling.download_progress — Method

download_progress(total, now; filename="")

ClimaOcean.DataWrangling.inpaint_mask! — Method

inpaint_mask!(field, mask; max_iter = Inf)

Inpaint field within mask, using values outside mask. In other words, regions where mask[i, j, k] == 1 will be inpainted and regions where mask[i, j, k] == 0 will be preserved.

Arguments

- `field`: `Field` to be inpainted.
- `mask`: Boolean-valued `Field`, values where
          `mask[i, j, k] == true` are inpainted.
- `max_iter`: Maximum iterations for inpainting. Non-Inf values mean that
              NaN's can occur within the mask.

ClimaOcean.DataWrangling.propagate_horizontally! — Function

propagate_horizontally!(field, mask [, tmp_field=deepcopy(field)]; max_iter = Inf)

Horizontally propagate the values of field into the mask. In other words, cells where mask[i, j, k] == false are preserved, and cells where mask[i, j, k] == true are painted over.

ECCO2

ClimaOcean.DataWrangling.ECCO2.inpainted_ecco2_field — Method

inpainted_ecco2_field(variable_name; 
                      architecture = CPU(),
                      filename = "./inpainted_ecco2_fields.nc",
                      mask = ecco2_center_mask(architecture))

Retrieve the ECCO2 field corresponding to variable_name inpainted to fill all the missing values in the original dataset.

Arguments:

variable_name: the variable name corresponding to the Dataset.

Keyword Arguments:

architecture: either CPU() or GPU().

filename: the path where to retrieve the data from. If the file does not exist, the data will be retrived from the ECCO2 dataset, inpainted, and saved to filename.

mask: the mask used to inpaint the field (see inpaint_mask!).

Bathymetry