Private types and functions

Documentation for ClimaOcean.jl's internal interface.

ClimaOcean

Diagnostics

InitialConditions

DataWrangling

ClimaOcean.DataWrangling.continue_downwards! — Method

continue_downwards!(field, mask)

Continue downwards a field with missing values within mask. Cells where mask[i, k, k] == false will be preserved.

ClimaOcean.DataWrangling.download_progress — Method

download_progress(total, now; filename="")

ClimaOcean.DataWrangling.inpaint_mask! — Method

inpaint_mask!(field, mask; max_iter = Inf)

Inpaint field within mask, using values outside mask. In other words, regions where mask[i, j, k] == 1 will be inpainted and regions where mask[i, j, k] == 0 will be preserved.

Arguments

- `field`: `Field` to be inpainted.
- `mask`: Boolean-valued `Field`, values where
          `mask[i, j, k] == true` are inpainted.
- `max_iter`: Maximum iterations for inpainting. Non-Inf values mean that
              NaN's can occur within the mask.

ClimaOcean.DataWrangling.propagate_horizontally! — Function

propagate_horizontally!(field, mask [, tmp_field=deepcopy(field)]; max_iter = Inf)

Horizontally propagate the values of field into the mask. In other words, cells where mask[i, j, k] == false are preserved, and cells where mask[i, j, k] == true are painted over.

ECCO

ClimaOcean.DataWrangling.ECCO.ECCONetCDFBackend — Method

ECCONetCDFBackend(length)

Represents an ECCO FieldTimeSeries backed by ECCO native .nc files. Each time instance is stored in an individual file.

ClimaOcean.DataWrangling.ECCO.ECCORestoring — Type

struct ECCORestoring{FTS, G, M, V, N} <: Function

A struct representing ECCO restoring.

Fields

ECCO_fts: The ECCO FTS on the native ECCO grid.

ECCO_grid: The native ECCO grid to interpolate from.

mask: A mask (could be a number, an array, a function or a field).

variable_name: The variable name of the variable that needs restoring.

λ⁻¹: The reciprocal of the restoring timescale.

ClimaOcean.DataWrangling.ECCO.ECCO_field_time_series — Method

ECCO_field_time_series(metadata::ECCOMetadata;
                       architecture = CPU(),
                       time_indices_in_memory = 2,
                       time_indexing = Cyclical(),
                       grid = nothing)

Create a field time series object for ECCO data.

Arguments:

metadata: An ECCOMetadata object containing information about the ECCO dataset.

Keyword Arguments:

architecture: The architecture to use for computations (default: CPU()).

timeindicesin_memory: The number of time indices to keep in memory (default: 2).

time_indexing: The time indexing scheme to use (default: Cyclical()).

grid: if not a nothing, the ECCO data is directly interpolated on the grid,

ClimaOcean.DataWrangling.ECCO.ECCO_times — Method

ECCO_times(metadata; start_time = metadata.dates[1])

Extracts the time values from the given metadata and calculates the time difference from the start time.

Arguments

metadata: The metadata containing the date information.

start_time: The start time for calculating the time difference. Defaults to the first date in the metadata.

Returns

An array of time differences in seconds.

ClimaOcean.DataWrangling.ECCO.download_dataset! — Method

download_dataset!(metadata::ECCOMetadata)

Download the dataset specified by the given metadata. If the metadata contains a single date, the dataset is downloaded directly. If the metadata contains multiple dates, the dataset is downloaded for each date individually. The data download requires a username and password to be provided in the ECCOUSERNAME and ECCOPASSWORD environment variables. This can be done by exporting the environment variables in the shell before running the script, or by launching julia with

ECCOUSERNAME=myuser ECCOPASSWORD=mypasswrd julia

Arguments

metadata::ECCOMetadata: The metadata specifying the dataset to be downloaded.

ClimaOcean.DataWrangling.ECCO.inpainted_ECCO_field — Method

inpainted_ECCO_field(variable_name; 
                     architecture = CPU(),
                     filename = "./inpainted_ECCO_fields.nc",
                     mask = ECCO_mask(architecture))

Retrieve the ECCO field corresponding to variable_name inpainted to fill all the missing values in the original dataset.

Arguments:

variable_name: the variable name corresponding to the Dataset.

Keyword Arguments:

architecture: either CPU() or GPU().

filename: the path where to retrieve the data from. If the file does not exist, the data will be downloaded from the ECCO dataset.

mask: the mask used to inpaint the field (see inpaint_mask!).

maxiter: the maximum number of iterations to inpaint the field (see inpaint_mask!).

Bathymetry

ClimaOcean.Bathymetry.remove_minor_basins! — Method

remove_minor_basins!(z_data, keep_major_basins)

Remove independent basins from the bathymetry data stored in z_data by identifying connected regions below sea level. Basins are removed from smallest to largest until only keep_major_basins remain.

Arguments

z_data: A 2D array representing the bathymetry data.

keep_major_basins: The maximum number of connected regions to keep. Default is Inf, which means all connected regions are kept.

VerticalGrids

ClimaOcean.VerticalGrids.stretched_vertical_faces — Method

stretched_vertical_faces(; surface_layer_Δz = 5.0,
                           surface_layer_height = 100.0,
                           constant_bottom_spacing_depth = Inf,
                           maximum_Δz = Inf,
                           stretching = PowerLawStretching(1.02),
                           rounding_digits = 1,
                           depth = 5000)

Return an array of cell interfaces with surface_layer_Δz spacing in a surface layer of height surface_layer_height, and stretched according to the function stretching(Δz_above, z_above) down to depth. The interfaces extends from Lz = -z[1] to 0 = z[end], where Lz ≥ depth.

The grid spacing Δz is limited to be less than maximum_Δz. The grid is also uniformly-spaced below constant_bottom_spacing_depth.

rounding_digits controls the accuracy with which the grid face positions are saved.

OceanSeaIceModels

CrossRealFluxes

ClimaOcean.OceanSeaIceModels.CrossRealmFluxes.LatitudeDependentAlbedo — Type

LatitudeDependentAlbedo([FT::DataType=Float64]; diffuse = 0.069, direct = 0.011)

Constructs a LatitudeDependentAlbedo object. The albedo of the ocean surface is assumed to be a function of the latitude, obeying the following formula (Large and Yeager, 2009):

α(φ) = α.diffuse - α.direct * cos(2φ)

where φ is the latitude, α.diffuse is the diffuse albedo, and α_.irect is the direct albedo.

Arguments

===========

FT::DataType: The data type of the albedo values. Default is Float64.

Keyword Arguments

===================

diffuse: The diffuse albedo value. Default is 0.069.

direct: The direct albedo value. Default is 0.011.

ClimaOcean.OceanSeaIceModels.CrossRealmFluxes.Radiation — Type

Radiation([arch = CPU(), FT=Float64];
          ocean_emissivity = 0.97,
          sea_ice_emissivity = 1.0,
          ocean_albedo = TabulatedAlbedo(arch, FT),
          sea_ice_albedo = 0.7,
          stefan_boltzmann_constant = 5.67e-8)

Constructs a Radiation object that represents the radiation properties of the ocean and sea ice.

Arguments

===========

arch: The architecture of the system (default: CPU()).

FT: The floating-point type to use (default: Float64).

Keyword Arguments

===================

ocean_emissivity: The emissivity of the ocean surface (default: 0.97).

sea_ice_emissivity: The emissivity of the sea ice surface (default: 1.0).

ocean_albedo: The albedo of the ocean surface (default: LatitudeDependentAlbedo(FT)).

sea_ice_albedo: The albedo of the sea ice surface (default: 0.7).

stefan_boltzmann_constant: The Stefan-Boltzmann constant (default: 5.67e-8).

ClimaOcean.OceanSeaIceModels.CrossRealmFluxes.SimilarityTheoryTurbulentFluxes — Type

SimilarityTheoryTurbulentFluxes(FT::DataType = Float64;
                                gravitational_acceleration = default_gravitational_acceleration,
                                von_karman_constant = convert(FT, 0.4),
                                turbulent_prandtl_number = convert(FT, 1),
                                gustiness_parameter = convert(FT, 6.5),
                                stability_functions = default_stability_functions(FT),
                                thermodynamics_parameters = PATP(FT),
                                water_vapor_saturation = ClasiusClapyeronSaturation(),
                                water_mole_fraction = convert(FT, 0.98),
                                roughness_lengths = default_roughness_lengths(FT),
                                similarity_profile_type = LogarithmicSimilarityProfile(),
                                bulk_velocity = RelativeVelocity(),
                                tolerance = 1e-8,
                                maxiter = 100,
                                fields = nothing)

SimilarityTheoryTurbulentFluxes contains parameters and settings to calculate sea-air turbulent fluxes using Monin-Obukhov similarity theory.

Keyword Arguments

gravitational_acceleration: The gravitational acceleration (default: defaultgravitationalacceleration).

von_karman_constant: The von Karman constant (default: 0.4).

turbulent_prandtl_number: The turbulent Prandtl number (default: 1).

gustiness_parameter: The gustiness parameter that accounts for low wind speed areas (default: 6.5).

stability_functions: The stability functions. Default: defaultstabilityfunctions(FT) that follow the formulation of Edson et al (2013).

thermodynamics_parameters: The thermodynamics parameters used to calculate atmospheric stability and saturation pressure. Default: PATP, alias for PrescribedAtmosphereThermodynamicsParameters.

water_vapor_saturation: The water vapor saturation law. Default: ClasiusClapyeronSaturation() that follows the Clasius Clapyeron pressure formulation.

water_mole_fraction: The water mole fraction used to calculate the seawatersaturationspecific_humidity. Default: 0.98, the rest is assumed to be other substances such as chlorine, sodium sulfide and magnesium.

roughness_lengths: The roughness lengths used to calculate the characteristic scales for momentum, temperature and water vapor. Default: defaultroughnesslengths(FT), formulation taken from Edson et al (2013).

similarity_profile_type: The type of similarity profile used to relate the atmospheric state to the surface fluxes / characteristic scales.

bulk_velocity: The velocity used to calculate the characteristic scales. Default: RelativeVelocity() (difference between atmospheric and oceanic speed).

tolerance: The tolerance for convergence (default: 1e-8).

maxiter: The maximum number of iterations (default: 100).

fields: The fields to calculate (default: nothing).