Public Documentation

Documentation for ClimaOcean.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

ClimaOcean

Diagnostics

ClimaOcean.Diagnostics.MixedLayerDepthField — Method

MixedLayerDepthField(grid, tracers, buoyancy_model; Δb = 3e-4, field_kw...)

Return a reduced Field{Center, Center, Nothing} that represents mixed layer depth for model, based on a buoyancy differential criterion. The mixed layer depth is defined as the depth $h$ for which

\[b(z=0) - b(z=-h) = Δb\]

This criterion is solved by integrating downwards and linearly interpolating to find h, assuming that $b$ decreases with depth.

Keyword arguments

Δb: Buoyancy differential used to calculate mixed layer depth

field_kw: Keyword arguments passed to Field.

Example

h = MixedLayerDepth(model)
compute!(h) # compute mixed layer depth

InitialConditions

DataWrangling

ECCO

ClimaOcean.DataWrangling.ECCO.ECCOMetadata — Type

ECCOMetadata{D, V}

Metadata holding the ECCO dataset information:

name: The name of the dataset.

dates: The dates of the dataset, in a AbstractCFDateTime format.

version: The version of the dataset, could be ECCO2Monthly, ECCO2Daily, or ECCO4Monthly.

dir: The directory where the dataset is stored.

ClimaOcean.DataWrangling.ECCO.ECCOMetadata — Method

ECCOMetadata(name::Symbol; 
             dates = DateTimeProlepticGregorian(1993, 1, 1), 
             version = ECCO4Monthly(), 
             dir = download_ECCO_cache)

Constructs an ECCOMetadata object with the specified parameters.

Arguments

============

name::Symbol: The name of the metadata.

Keyword Arguments

===================

date: The date of the metadata (default: DateTimeProlepticGregorian(1993, 1, 1)).

version: The version of the metadata (for the moment the choices are ECCO2Monthly(), ECCO2Daily(), or ECCO4Monthly()).

dir: The dir to the datafile (default: downloadECCOcache).

ClimaOcean.DataWrangling.ECCO.ECCO_field — Method

ECCO_field(variable_name;
           architecture = CPU(),
           horizontal_halo = (1, 1),
           user_data = nothing,
           url = ecco_urls[variable_name],
           short_name = ecco_short_names[variable_name])

Retrieve the ECCO field corresponding to variable_name. The data is either: (1) retrieved from filename, (2) dowloaded from url if filename does not exists, (3) filled from user_data if user_data is provided.

ClimaOcean.DataWrangling.ECCO.ECCO_mask — Function

ECCO_mask(architecture = CPU(); minimum_value = Float32(-1e5))

A boolean field where true represents a missing value in the ECCO dataset.

ClimaOcean.DataWrangling.ECCO.ECCO_restoring_forcing — Function

ECCO_restoring_forcing(metadata::ECCOMetadata;
                       architecture = CPU(), 
                       backend = ECCONetCDFBackend(2),
                       time_indexing = Cyclical(),
                       mask = 1,
                       timescale = 5days)

Create a restoring forcing term that restores to values stored in an ECCO field time series.

Arguments:

=============

metadata: The metadata for the ECCO field time series.

Keyword Arguments:

====================

architecture: The architecture. Typically CPU or GPU

time_indices_in_memory: The number of time indices to keep in memory. trade-off between performance and memory footprint.

time_indexing: The time indexing scheme for the field time series, see FieldTimeSeries

mask: The mask value. Can be a function of (x, y, z, time), an array or a number

timescale: The restoring timescale.

Bathymetry

ClimaOcean.Bathymetry.regrid_bathymetry — Method

regrid_bathymetry(target_grid;
                  url = "https://www.ngdc.noaa.gov/thredds/fileServer/global/ETOPO2022/60s/60s_surface_elev_netcdf", 
                  height_above_water = <none>,
                  minimum_depth = 0,
                  dir = download_cache,
                  filename = "ETOPO_2022_v1_60s_N90W180_surface.nc")

Regrid bathymetry associated with the NetCDF file at path = joinpath(dir, filename) to target_grid. If path does not exist, then a download is attempted from joinpath(url, filename).

Arguments:

target_grid: grid to interpolate onto

Keyword Arguments:

heightabovewater: limits the maximum height of above-water topography (where h > 0). If nothing the original topography is retained

minimumdepth: minimum depth for the shallow regions, defined as a positive value. `h > - minimumdepth` will be considered land

dir: directory of the bathymetry-containing file

filename: file containing bathymetric data. Must be netcdf with fields: (1) lat vector of latitude nodes (2) lon vector of longitude nodes (3) z matrix of depth values

interpolationpasses: regridding/interpolation passes. The bathymetry is interpolated in `interpolationpasses - 1` intermediate steps. With more steps the final bathymetry will be smoother. Example: interpolating from a 400x200 grid to a 100x100 grid in 4 passes will involve - 400x200 -> 325x175 - 325x175 -> 250x150 - 250x150 -> 175x125 - 175x125 -> 100x100 If coarsening the original grid, linear interpolation in passes is equivalent to applying a smoothing filter, with more passes increasing the strength of the filter. If refining the original grid, additional passes will not help and no intermediate steps will be performed.

majorbasins: Number of "independent major basins", or fluid regions fully encompassed by land, that are retained by [`removeminorbasins!](@ref). Basins are removed by order of size: the smallest basins are removed first.majorbasins=1will retain only the largest basin. Default:Inf`, which does not remove any basins.

ClimaOcean.Bathymetry.retrieve_bathymetry — Method

retrieve_bathymetry(grid, filename; kw...)

Retrieve the bathymetry data from a file or generate it using a grid and save it to a file.

Arguments

============

grid: The grid used to generate the bathymetry data.

filename: The name of the file to read or save the bathymetry data.

kw...: Additional keyword arguments.

Returns

===========

bottom_height: The retrieved or generated bathymetry data.

If the specified file exists, the function reads the bathymetry data from the file. Otherwise, it generates the bathymetry data using the provided grid and saves it to the file before returning it.